Move auth logic from the route to the controller
