+++ server/routes/auth.route.js
@@ -5,38 +5,8 @@ import paramValidation from '../../config/param-validation';
 import authCtrl from '../controllers/auth.controller';
 import config from '../../config/config';
 import passport from 'passport';
-import FacebookTokenStrategy from 'passport-facebook-token';
 
 const router = express.Router(); // eslint-disable-line new-cap
-
-
-passport.serializeUser(function(user, done) {
-  done(null, user);
-});
-
-passport.deserializeUser(function(user, done) {
-  done(null, user);
-});
-
-passport.use(new FacebookTokenStrategy({
-    clientID: config.facebook.clientID,
-    clientSecret: config.facebook.clientSecret
-  },
-  function(accessToken, refreshToken, profile, done) {
-    const user = {
-        'email': profile.emails[0].value,
-        'name' : profile.name.givenName + ' ' + profile.name.familyName,
-        'id'   : profile.id,
-        'token': accessToken
-    }
-
-    // You can perform any necessary actions with your user at this point,
-    // e.g. internal verification against a users table,
-    // creating new user entries, etc.
-
-    return done(null, user); // the user object we just made gets passed to the route's controller as `req.user`
-  }
-));
 
 /** POST /api/auth/login - Returns token if correct username and password is provided */
 router.route('/login')
